Add ParameterBounds and ParameterSetting test for Delay

+++ src/lib_DelayTankEngine/include/Delay.h
@@ -13,7 +13,7 @@ public:
 
 	void setDelay(float delayInSeconds);
 	void setGain(float gain);
-	void setPan(int pan);
+	void setPan(float pan);
 	std::pair<float, float> process(float input);
 
 	float getDelay() const;
@@ -22,7 +22,7 @@ public:
 private:
 	std::unique_ptr<CRingBuffer<float>> mDelayLine;
 
-	int mPan = 0;
+	float mPan = 0.0f;
 	float mGain = 1.0f;
 	float mSampleRate = 44100.0f;
 };
